+++ src/app/page.tsx
@@ -2,7 +2,7 @@ import { SignInButton } from "@clerk/nextjs";
 
 export default function HomePage() {
   return (
-    <SignInButton mode="redirect" redirectTo="/Home">
+    <SignInButton mode="redirect" forceRedirectUrl="/Home">
       <button className="btn btn-primary">Sign In</button>
     </SignInButton>
   );
